requirements.txt
langchain 
streamlit
torch
accelerate
sentence_transformers
python-dotenv
huggingface_hub
ctransformers
replicate